[backend] support newer versions of lxc

+++ src/backend/call-service-in-lxc.sh
@@ -1,9 +1,13 @@
 #!/bin/bash
 
-#FSDIR="/opt/obs/SourceServiceSystem"
-FSDIR="/"
+#set -x
+
+FSDIR="/opt/obs/SourceServiceSystem"
 MOUNTDIR="/opt/obs/SourceServiceSystem.mounts/"
 TEMPDIR="/lxc.tmp.$$"
+RETURN="0"
+
+# set -x
 
 INNEROUTDIR="$TEMPDIR/out"
 INNERSRCDIR="$TEMPDIR/src"
@@ -35,7 +39,7 @@ WITH_NET="0"
 COMMAND="$1"
 shift
 case "$COMMAND" in
-  */download_url|*/tar_scm|*/download_src_package|*/update_source|*/download_files)
+  */download_url|*/tar_scm|*/download_src_package|*/update_source|*/download_files|*/generator_pom)
     WITH_NET="1"
     ;;
 esac
@@ -62,13 +66,17 @@ chown -R $RUNUSER "$MOUNTDIR/$INNEROUTDIR"
 #  echo "rcnscd start" >> "$MOUNTDIR/$INNERSCRIPT"
 #fi
 echo -n "su $RUNUSER -s ${INNERSCRIPT}.command" >> "$MOUNTDIR/$INNERSCRIPT"
-echo "#!/bin/bash"                         >  "$MOUNTDIR/${INNERSCRIPT}.command"
+echo "#!/bin/bash"               >  "$MOUNTDIR/${INNERSCRIPT}.command"
+#echo "set -x" >> "$MOUNTDIR/${INNERSCRIPT}.command"
+#echo "ls -ld /dev /dev/null" >> "$MOUNTDIR/${INNERSCRIPT}.command"
+echo "echo Running ${COMMAND[@]} --outdir $INNEROUTDIR" >> "$MOUNTDIR/${INNERSCRIPT}.command"
 echo "${COMMAND[@]} --outdir $INNEROUTDIR" >> "$MOUNTDIR/${INNERSCRIPT}.command"
 chmod 0755 "$MOUNTDIR/$INNERSCRIPT" "$MOUNTDIR/${INNERSCRIPT}.command"
 
 # construct jail
 LXC_CONF="/obs.service.$$"
 echo "lxc.utsname = obs.service.$$" > $LXC_CONF
+mount -t proc proc $MOUNTDIR/proc
 if [ "$WITH_NET" != "1" ] ; then
   echo "lxc.network.type = empty" >> $LXC_CONF
   echo "lxc.network.flags = up" >> $LXC_CONF
@@ -77,36 +85,46 @@ fi
 echo "lxc.tty = 1" >> $LXC_CONF
 #echo "lxc.mount = /etc/fstab" >> $LXC_CONF
 echo "lxc.rootfs = $MOUNTDIR" >> $LXC_CONF
-mount -t proc proc $MOUNTDIR/proc
+echo "lxc.autodev = 1" >> $LXC_CONF
+echo "lxc.cgroup.devices.allow = c 1:3 rw" >> $LXC_CONF
 
 lxc-info -n obs.service.jail.$$ >& /dev/null && lxc-destroy -n obs.service.jail.$$ >& /dev/null
 RETURN="0"
 lxc-create -n obs.service.jail.$$ -f $LXC_CONF >& /dev/null || RETURN="2"
-rm -f $LXC_CONF
 
 # run jailed process
-lxc-start -n obs.service.jail.$$ "$INNERSCRIPT" || RETURN="2"
-
-# destroy jail
-lxc-destroy -n obs.service.jail.$$
-
-# move out the result
-if [ 0`find "$MOUNTDIR/$INNEROUTDIR" -type f | wc -l` -gt 0 ]; then
-  for i in _service:* ; do
-    if [ ! -f "$MOUNTDIR/$INNERSRCDIR/$i" ]; then
-      rm -f "$i"
-    fi
-  done
+if lxc-start -n obs.service.jail.$$ "$INNERSCRIPT"; then
+  # move out the result
+  if [ 0`find "$MOUNTDIR/$INNEROUTDIR" -type f | wc -l` -gt 0 ]; then
+    for i in _service:* ; do
+      if [ ! -f "$MOUNTDIR/$INNERSRCDIR/$i" ]; then
+        rm -f "$i"
+      fi
+    done
+  fi
+else
+ RETURN="2"
 fi
+
+#ls $FSDIR
 
 # cleanup
 umount "$MOUNTDIR/proc"
 umount "$MOUNTDIR$INNERSRCDIR"
 umount "$MOUNTDIR$INNEROUTDIR"
-umount "$MOUNTDIR"
-rm "$MOUNTDIR/$INNERSCRIPT"
-rmdir "$MOUNTDIR/$TEMPDIR"
-rmdir "$MOUNTDIR"
+rmdir --ignore-fail-on-non-empty "$MOUNTDIR/$INNERSRCDIR"
+rmdir --ignore-fail-on-non-empty "$MOUNTDIR/$INNEROUTDIR"
+rm -f "$MOUNTDIR/$INNERSCRIPT.command" 2> /dev/null
+rm -f "$MOUNTDIR/$INNERSCRIPT" 2> /dev/null
+rmdir --ignore-fail-on-non-empty "$MOUNTDIR/$TEMPDIR" 2> /dev/null
+umount "$MOUNTDIR" 
+rmdir --ignore-fail-on-non-empty "$MOUNTDIR" 2> /dev/null
+#ls $FSDIR
+
+# destroy jail
+# lxc-destroy -n obs.service.jail.$$
+# lxc-destory removes the entire system now
+rm -rf /var/lib/lxc/obs.service.jail.$$
 
 exit $RETURN
